scoreboard updates and game score updates

+++ minigamemanager/src/main/java/io/zkz/mc/minigameplugins/minigamemanager/score/ScoreEntry.java
@@ -2,5 +2,5 @@ package io.zkz.mc.minigameplugins.minigamemanager.score;
 
 import java.util.UUID;
 
-public record ScoreEntry(UUID playerId, String minigame, int round, String reason, float points) {
+public record ScoreEntry(UUID playerId, String minigame, int round, String reason, double points) {
 }
